--- src/main/java/com/northcoders/makemydayapi/service/ActivityServiceImpl.java
@@ -1,4 +1,0 @@
-package com.northcoders.makemydayapi.service;
-
-public class ActivityServiceImpl implements ActivityService {
-}
