--- src/main/java/com/northcoders/makemydayapi/service/GooglePlacesServiceImpl.java
@@ -1,4 +1,0 @@
-package com.northcoders.makemydayapi.service;
-
-public class GooglePlacesServiceImpl implements GooglePlacesService {
-}
