--- src/main/java/com/northcoders/makemydayapi/service/TicketmasterServiceImpl.java
@@ -1,14 +1,0 @@
-package com.northcoders.makemydayapi.service;
-
-import com.northcoders.makemydayapi.model.Activity;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-
-@Service
-public class TicketmasterServiceImpl implements TicketmasterService {
-    @Override
-    public List<Activity> getAllEvents() {
-        return List.of();
-    }
-}
