+++ src/main/java/com/northcoders/makemydayapi/service/impl/ActivityServiceImpl.java
@@ -1,0 +1,6 @@
+package com.northcoders.makemydayapi.service.impl;
+
+import com.northcoders.makemydayapi.service.ActivityService;
+
+public class ActivityServiceImpl implements ActivityService {
+}
