+++ src/main/java/com/northcoders/makemydayapi/service/impl/GooglePlacesServiceImpl.java
@@ -1,0 +1,6 @@
+package com.northcoders.makemydayapi.service.impl;
+
+import com.northcoders.makemydayapi.service.GooglePlacesService;
+
+public class GooglePlacesServiceImpl implements GooglePlacesService {
+}
