+++ src/main/java/com/northcoders/makemydayapi/service/impl/TicketmasterServiceImpl.java
@@ -1,0 +1,15 @@
+package com.northcoders.makemydayapi.service.impl;
+
+import com.northcoders.makemydayapi.model.Activity;
+import com.northcoders.makemydayapi.service.TicketmasterService;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public class TicketmasterServiceImpl implements TicketmasterService {
+    @Override
+    public List<Activity> getAllEvents() {
+        return List.of();
+    }
+}
